requirements.txt
pendulum==3.1.0
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.2
